+++ src/components/script/ScriptEditorEvent.tsx
@@ -226,6 +226,7 @@ const ScriptEditorEvent = React.memo(
       collect: (monitor: DragSourceMonitor) => ({
         isDragging: monitor.isDragging(),
       }),
+      canDrag: () => !rename,
     });
 
     const toggleOpen = useCallback(
@@ -289,6 +290,7 @@ const ScriptEditorEvent = React.memo(
     const isOpen = (scriptEvent?.args && !scriptEvent.args.__collapse) ?? false;
     const isConditional = scriptEventDefs[command]?.isConditional ?? false;
     const editableSymbol = scriptEventDefs[command]?.editableSymbol ?? false;
+    const description = scriptEventDefs[command]?.description;
 
     const onOpenAddMenu = useCallback((before: boolean) => {
       setInsertBefore(before);
@@ -502,7 +504,7 @@ const ScriptEditorEvent = React.memo(
 
         <div ref={dropRef}>
           {isOverCurrent && <ScriptEventPlaceholder />}
-          <div ref={dragRef}>
+          <div ref={dragRef} title={description}>
             <ScriptEventHeader
               ref={headerRef}
               isConditional={isConditional}