+++ src/lib/project/migration/migrateProjectResources.ts
@@ -12,6 +12,7 @@ import {
   migrate420r5To420r6,
   migrate420r6To420r7,
   migrate420r7To420r8,
+  migrate420r8To420r9,
 } from "./versions/410to420";
 
 const migrations: ProjectResourcesMigration[] = [
@@ -24,6 +25,7 @@ const migrations: ProjectResourcesMigration[] = [
   migrate420r5To420r6,
   migrate420r6To420r7,
   migrate420r7To420r8,
+  migrate420r8To420r9,
 ];
 
 const lastMigration = migrations[migrations.length - 1];